Rewrite of integration tests using pytest instead of unittest.TestCase.
In preparation for the modification of the integration tests to cover
GSSAPI authentication I had to make a few changes to the existing
test cases, which were written using unittest.TestCase.

After some discussions with @dpkp and @jeffwidman I decided to first
move those test cases to pytest, which is the preferred test framework
for the project.